requirements.txt
Flask
google-cloud-pubsub
psycopg2-binary
SQLAlchemy
python-dotenv 